pyproject.toml
[tool.poetry]
name = "espanso-edit"
version = "0.1.0"
description = ""
authors = ["PranitSh <ppshah2023@gmail.com>"]
readme = "README.md"
packages = [{include = "espanso_edit"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
easygui = "^0.98.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
